docs/{{ cookiecutter.name }}/{{ cookiecutter.name }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.name }}"
version = "0.1.0"
description = "{{ cookiecutter.friendly_name }} stream processing pipeline application for Tradesignals.io."
authors = ["Dev Ops <devops@tradesignals.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
tornado = "^6.4.2"
confluent-kafka = {extras = ["avro"], version = "^2.6.1"}
attrs = "^24.2.0"
cattrs = "^24.1.2"
pandas-market-calendars = "^4.4.2"
orjson = "^3.10.12"
pyyaml = "^6.0.2"
pandas = "^2.2.3"
uvloop = "^0.21.0"
rich = "^13.9.4"
databento = "^0.46.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
start-pipeline = "{{ cookiecutter.name }}.pipeline:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
